+++ harry_potter/src/data/index.ts
@@ -1,4 +1,2 @@
-export { initialState } from './initData.tsx';
-export { potterApi } from './apiRequest.tsx';
 export type { RootState } from './store.tsx';
 export { store } from './store.tsx';
