--- harry_potter/src/data/initData.tsx
@@ -1,20 +1,0 @@
-import { State } from '../types/types';
-
-export const initialState: State = {
-  searchTerm: localStorage.getItem('searchTerm') || '',
-  charactersList: [],
-  selectedCharacters: [],
-  loading: true,
-  error: {
-    message: '',
-    stack: '',
-  },
-  showErrorModal: false,
-  errorThrow: false,
-  detailesOpened: false,
-  pageNumber: 1,
-  pageSize: 30,
-  records: 0,
-  detailedId: '',
-  detailedCard: undefined,
-};
